requirements/base.txt
# Base dependencies
# Django maybe missing here if multiple versions are to be tested and installed by tox
# For development we keep them unfreezed to work and test on newest versions as soon as they are released
# Freezing can be performed by pip-compile or pipenv

click
click-log
colorlog
crayons
dinja2
# https://docs.djangoproject.com/en/dev/faq/install/#faq-python-version-support
django<2.0 ; python_version < '3.4'
django<2.1 ; python_version == '3.4'
django     ; python_version >= '3.5'
django-admin-ordering
django-admin-relation-links
django-assets
django-bootstrap-form
django-babel
django-debug-toolbar
django-email-queue
django-environ
django-error-views
django-formtools
django-import-export
# django-localflavor
django-opt-out[sparkpost]
django-pascal-templates
# django-polymodels
django-powerbank
django-pure-pagination
django-reversion
# easy_thumbnails
djangorestframework
djangorestframework-jsonapi
enum34 ; python_version < '3.4'
html2text
# isoweek
# jinja2-highlight
# markdown
invoke
# maya
newrelic
# opencage
psycopg2-binary
pycountry
pytz
pathlib ; python_version < '3.3'
pendulum
python-dotenv
# version 2018.11.3 has installation issues
# regex==2018.8.29
# sorl-thumbnail
raven
requests
whitenoise

--find-links=local

